refactor quay registry function
