+++ lib/datasource/docker/common.ts
@@ -1,0 +1,146 @@
+import { ECR, ECRClientConfig } from '@aws-sdk/client-ecr';
+import wwwAuthenticate from 'www-authenticate';
+import { HOST_DISABLED } from '../../constants/error-messages';
+import { logger } from '../../logger';
+import { HostRule } from '../../types';
+import { ExternalHostError } from '../../types/errors/external-host-error';
+import * as hostRules from '../../util/host-rules';
+import { Http } from '../../util/http';
+import type { OutgoingHttpHeaders } from '../../util/http/types';
+
+export const ecrRegex = /\d+\.dkr\.ecr\.([-a-z0-9]+)\.amazonaws\.com/;
+
+export async function getECRAuthToken(
+  region: string,
+  opts: HostRule
+): Promise<string | null> {
+  const config: ECRClientConfig = { region };
+  if (opts.username && opts.password) {
+    config.credentials = {
+      accessKeyId: opts.username,
+      secretAccessKey: opts.password,
+    };
+  }
+  const ecr = new ECR(config);
+  try {
+    const data = await ecr.getAuthorizationToken({});
+    const authorizationToken = data?.authorizationData?.[0]?.authorizationToken;
+    if (authorizationToken) {
+      return authorizationToken;
+    }
+    logger.warn(
+      'Could not extract authorizationToken from ECR getAuthorizationToken response'
+    );
+  } catch (err) {
+    logger.trace({ err }, 'err');
+    logger.debug('ECR getAuthorizationToken error');
+  }
+  return null;
+}
+
+export async function getAuthHeaders(
+  id: string,
+  http: Http,
+  registry: string,
+  dockerRepository: string
+): Promise<OutgoingHttpHeaders | null> {
+  try {
+    const apiCheckUrl = `${registry}/v2/`;
+    const apiCheckResponse = await http.get(apiCheckUrl, {
+      throwHttpErrors: false,
+    });
+    if (apiCheckResponse.headers['www-authenticate'] === undefined) {
+      return {};
+    }
+    const authenticateHeader = new wwwAuthenticate.parsers.WWW_Authenticate(
+      apiCheckResponse.headers['www-authenticate']
+    );
+
+    const opts: HostRule & {
+      headers?: Record<string, string>;
+    } = hostRules.find({ hostType: id, url: apiCheckUrl });
+    if (ecrRegex.test(registry)) {
+      const [, region] = ecrRegex.exec(registry);
+      const auth = await getECRAuthToken(region, opts);
+      if (auth) {
+        opts.headers = { authorization: `Basic ${auth}` };
+      }
+    } else if (opts.username && opts.password) {
+      const auth = Buffer.from(`${opts.username}:${opts.password}`).toString(
+        'base64'
+      );
+      opts.headers = { authorization: `Basic ${auth}` };
+    }
+    delete opts.username;
+    delete opts.password;
+
+    if (authenticateHeader.scheme.toUpperCase() === 'BASIC') {
+      logger.debug(`Using Basic auth for docker registry ${dockerRepository}`);
+      await http.get(apiCheckUrl, opts);
+      return opts.headers;
+    }
+
+    // prettier-ignore
+    const authUrl = `${String(authenticateHeader.parms.realm)}?service=${String(authenticateHeader.parms.service)}&scope=repository:${dockerRepository}:pull`;
+    logger.trace(
+      `Obtaining docker registry token for ${dockerRepository} using url ${authUrl}`
+    );
+    const authResponse = (
+      await http.getJson<{ token?: string; access_token?: string }>(
+        authUrl,
+        opts
+      )
+    ).body;
+
+    const token = authResponse.token || authResponse.access_token;
+    // istanbul ignore if
+    if (!token) {
+      logger.warn('Failed to obtain docker registry token');
+      return null;
+    }
+    return {
+      authorization: `Bearer ${token}`,
+    };
+  } catch (err) /* istanbul ignore next */ {
+    if (err.host === 'quay.io') {
+      // TODO: debug why quay throws errors (#9604)
+      return null;
+    }
+    if (err.statusCode === 401) {
+      logger.debug(
+        { registry, dockerRepository },
+        'Unauthorized docker lookup'
+      );
+      logger.debug({ err });
+      return null;
+    }
+    if (err.statusCode === 403) {
+      logger.debug(
+        { registry, dockerRepository },
+        'Not allowed to access docker registry'
+      );
+      logger.debug({ err });
+      return null;
+    }
+    // prettier-ignore
+    if (err.name === 'RequestError' && registry.endsWith('docker.io')) { // lgtm [js/incomplete-url-substring-sanitization]
+        throw new ExternalHostError(err);
+      }
+    // prettier-ignore
+    if (err.statusCode === 429 && registry.endsWith('docker.io')) { // lgtm [js/incomplete-url-substring-sanitization]
+        throw new ExternalHostError(err);
+      }
+    if (err.statusCode >= 500 && err.statusCode < 600) {
+      throw new ExternalHostError(err);
+    }
+    if (err.message === HOST_DISABLED) {
+      logger.trace({ registry, dockerRepository, err }, 'Host disabled');
+      return null;
+    }
+    logger.warn(
+      { registry, dockerRepository, err },
+      'Error obtaining docker token'
+    );
+    return null;
+  }
+}
